Add Rx module
