+++ core/.gitignore
