+++ core/bintray.gradle
@@ -1,0 +1,49 @@
+apply plugin: 'maven-publish'
+apply plugin: 'com.jfrog.bintray'
+
+publishing {
+    publications {
+        Publication(MavenPublication) {
+            artifact("$buildDir/outputs/aar/core-release.aar")
+            groupId 'co.infinum'
+            artifact sourcesJar
+            artifact javadocsJar
+            artifactId 'goldfinger'
+            version versions.goldfinger
+
+            pom.withXml {
+                def root = asNode()
+                def dependenciesNode = root.appendNode('dependencies')
+
+                // Iterate over the implementation dependencies (we don't want the test ones), adding a <dependency> node for each
+                configurations.implementation.allDependencies.each {
+                    // Ensure dependencies such as fileTree are not included in the pom.
+                    if (it.name != 'unspecified') {
+                        def dependencyNode = dependenciesNode.appendNode('dependency')
+                        dependencyNode.appendNode('groupId', it.group)
+                        dependencyNode.appendNode('artifactId', it.name)
+                        dependencyNode.appendNode('version', it.version)
+                    }
+                }
+            }
+        }
+    }
+}
+
+bintray {
+    user = System.getenv('BINTRAY_USER')
+    key = System.getenv('BINTRAY_API_KEY')
+    publish = false
+    publications = ['Publication']
+
+    pkg {
+        repo = 'android'
+        name = 'goldfinger'
+        userOrg = 'infinum'
+        publicDownloadNumbers = true
+
+        version {
+            name = versions.goldfinger
+        }
+    }
+}
