+++ core/src/main/java/co/infinum/goldfinger/Clock.java
