+++ core/src/main/java/co/infinum/goldfinger/CryptoObjectInitRunnable.java
