+++ core/src/main/java/co/infinum/goldfinger/Mode.java
