+++ example/src/main/AndroidManifest.xml
@@ -7,12 +7,16 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".ExampleActivity">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
+
+        <activity android:name=".ExampleActivity" android:screenOrientation="portrait"/>
+        <activity android:name=".RxExampleActivity" android:screenOrientation="portrait"/>
+
     </application>
 
 </manifest>
