--- lib/build.gradle
@@ -1,23 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion sdk.compile
-
-    defaultConfig {
-        minSdkVersion sdk.min
-        targetSdkVersion sdk.target
-        versionName versions.goldfinger
-    }
-
-    testOptions {
-        unitTests.returnDefaultValues = true
-    }
-}
-
-dependencies {
-    implementation "com.android.support:appcompat-v7:${versions.google}"
-    testImplementation 'junit:junit:4.12'
-    testImplementation 'org.mockito:mockito-core:2.10.0'
-}
-
-apply from: 'bintray.gradle'
