+++ rx/src/test/java/co/infinum/goldfinger/rx/TestObserver.java
@@ -1,0 +1,18 @@
+package co.infinum.goldfinger.rx;
+
+import io.reactivex.observers.DisposableObserver;
+
+public class TestObserver extends DisposableObserver<GoldfingerEvent> {
+
+    @Override
+    public void onComplete() {
+    }
+
+    @Override
+    public void onError(Throwable e) {
+    }
+
+    @Override
+    public void onNext(GoldfingerEvent goldfingerEvent) {
+    }
+}
