+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':lib', ':example'
+include ':core', ':example', ':rx'
